+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/ChatCell/ChatBaseCell.swift
@@ -107,8 +107,6 @@ class ChatBaseCell: UITableViewCell {
       .systemRed.withAlphaComponent(0.1)
     case .loading:
       .systemGray6
-    case .attachment:
-      .systemPurple.withAlphaComponent(0.1)
     }
   }
 }
