--- packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/ChatCell/ViewModel/AttachmentCellViewModel.swift
@@ -1,23 +1,0 @@
-//
-//  AttachmentCellViewModel.swift
-//  Intelligents
-//
-//  Created by 秋星桥 on 6/26/25.
-//
-
-import Foundation
-
-struct AttachmentCellViewModel: ChatCellViewModel {
-  var cellType: CellType = .attachment
-  var id: UUID
-  var attachments: [AttachmentViewModel]
-  var parentMessageId: String
-}
-
-struct AttachmentViewModel: Codable, Identifiable, Equatable, Hashable {
-  var id: String
-  var url: String
-  var mimeType: String?
-  var fileName: String?
-  var size: Int64?
-}
