+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/ChatCell/ViewModel/CellType.swift
@@ -11,7 +11,6 @@ public enum CellType: String, Codable, CaseIterable {
   case userMessage
   case assistantMessage
   case systemMessage
-  case attachment
   case loading
   case error
 }
