+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/ChatCell/ViewModel/ErrorCellViewModel.swift
@@ -11,6 +11,4 @@ struct ErrorCellViewModel: ChatCellViewModel {
   var cellType: CellType = .error
   var id: UUID
   var errorMessage: String
-  var canRetry: Bool
-  var retryAction: String?
 }
